refactor(types): tags-input
